+++ test/c99/test-errors.c
@@ -843,6 +843,7 @@ static int test_bary2obs() {
 }
 
 static int test_geo_posvel() {
+  enum novas_debug_mode m = novas_get_debug_mode();
   observer o;
   double p[3] = {0.0}, v[3] = {0.0};
   int n = 0;
@@ -851,6 +852,10 @@ static int test_geo_posvel() {
   if(check("geo_posvel:loc", -1, geo_posvel(0.0, 0.0, NOVAS_FULL_ACCURACY, NULL, p, v))) n++;
   if(check("geo_posvel:pos+vel", -1, geo_posvel(0.0, 0.0, NOVAS_FULL_ACCURACY, &o, NULL, NULL))) n++;
   if(check("geo_posvel:accuracy", 1, geo_posvel(0.0, 0.0, -1, &o, p, v))) n++;
+
+  novas_debug(NOVAS_DEBUG_EXTRA);
+  if(check("geo_posvel:geodetic", -1, geo_posvel(0.0, 0.0, NOVAS_REDUCED_ACCURACY, &o, p, v))) n++;
+  novas_debug(m);
 
   o.where = -1;
   if(check("geo_posvel:where", 2, geo_posvel(0.0, 0.0, NOVAS_FULL_ACCURACY, &o, p, v))) n++;
@@ -2640,8 +2645,6 @@ static int test_moon_elp_sky_pos() {
   return n;
 }
 
-
-
 static int test_uvw() {
   int n = 0;
 
@@ -2663,6 +2666,21 @@ static int test_site_uvw() {
 
   if(check("site_uvw:time:NULL", -1, novas_site_uvw(NULL, &s, p, 0.0, 0.0, NOVAS_REDUCED_ACCURACY, u))) n++;
   if(check("site_uvw:station:NULL", -1, novas_site_uvw(&ts, NULL, p, 0.0, 0.0, NOVAS_REDUCED_ACCURACY, u))) n++;
+
+  return n;
+}
+
+static int test_site_gcrs_posvel() {
+  int n = 0;
+
+  novas_timespec ts = {};
+  on_surface s = {};
+  double p[3] = {0.0}, v[3] = {0.0}, vs[3] = {0.0};
+
+  if(check("site_gcrs_posvel:ts:NULL", -1, novas_site_gcrs_posvel(NULL, &s, vs, 0.0, 0.0, NOVAS_REDUCED_ACCURACY, p, v))) n++;
+  if(check("site_gcrs_posvel:site:NULL", -1, novas_site_gcrs_posvel(&ts, NULL, vs, 0.0, 0.0, NOVAS_REDUCED_ACCURACY, p, v))) n++;
+  if(check("site_gcrs_posvel:accuracy:-1", -1, novas_site_gcrs_posvel(&ts, &s, vs, 0.0, 0.0, (enum novas_accuracy) -1, p, v))) n++;
+  if(check("site_gcrs_posvel:accuracy:p=v=NULL", -1, novas_site_gcrs_posvel(&ts, &s, vs, 0.0, 0.0, NOVAS_REDUCED_ACCURACY, NULL, NULL))) n++;
 
   return n;
 }
@@ -2897,6 +2915,7 @@ int main(int argc, const char *argv[]) {
 
   if(test_uvw()) n++;
   if(test_site_uvw()) n++;
+  if(test_site_gcrs_posvel()) n++;
 
   if(n) fprintf(stderr, " -- FAILED %d tests\n", n);
   else fprintf(stderr, " -- OK\n");