+++ CHANGELOG.md
@@ -32,6 +32,9 @@ Upcoming feature release, introducing a proper C++ API for the first time, and b
    
  - #298: Added `cspice_clear_kernels()` to CSPICE plugin to close all kernels and free up the resources they use 
    (thanks to aleberti)
+   
+ - #306: Added `novas_geodetic_posvel()` as a more precise alternative to `geo_posvel()` for Earth-bound observers,
+   when polar wobble offsets are known.
 
  - Added `novas_enu_to_itrs()` and `novas_itrs_to_enu()` functions to help convert between local East-North-Up (ENU)
    coordinates and ITRS. ENU is a natural local cartesian coordinate system of an observer at or near the Earth's 
@@ -43,6 +46,12 @@ Upcoming feature release, introducing a proper C++ API for the first time, and b
 ### Changed
 
  - #256: Changes to repo layout to accommodate C++ API _in addition to_ the base C99 API.
+ 
+ - #306: `novas_make_frame()` now uses the new `novas_site_gcrs_posvel()` when calculating SSB-based observer location 
+   and velocity in the GCRS.
+   
+ - #306: `geo_posvel()` changed to return an error (-1) if used for a geodetic observer and debug mode is set to
+   `NOVAS_DEBUG_EXTRA`, to warn that polar offsets are not included in the calculation. 
 
  - Use more precise matrix from Liu et al. (2011) for equatorial / galactic conversions.